A67
The program uses 3 swap functions that it overloads. The first swap function uses a temporary variable in order to switch two values that it has been provided with. When swap has 3 inputs, it checks the maximum, minimum, and middle value using couple if statements that compares if the number is larger, smaller, or in the middle of the previous number. When the swap function has 4 inputs, the functions swaps the values using the 2 input swap function 3 times. It does this in a specific way that moves all the variables left by a step and moves the first variable to the last spot.